Move things over to hints module
