+++ records_mover/db/mysql/load_options.py
@@ -2,7 +2,7 @@ from sqlalchemy.sql.expression import text, TextClause
 from records_mover.utils import quiet_remove
 from records_mover.records.hints import cant_handle_hint
 from records_mover.records import DelimitedRecordsFormat
-from records_mover.records.types import (
+from records_mover.records.hints import (
     HintEncoding, HintRecordTerminator,
     HintFieldDelimiter, HintQuoteChar,
     HintQuoting, HintEscape,
